requirements.txt
discord.py
lavalink
mutagen
python-dotenv
PyNaCl
typing_extensions
